requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
python-dotenv
markdown
Flask-Login
Flask-Bcrypt